{"schema":"decodex/commit/1","summary":"Modularize settings appearance controls","authority":"manual"}
